make ParserCalories more OOP-like, re-organised JUnit tests
